work on loader api
