+++ packages/connect/package.json
@@ -50,7 +50,7 @@
   },
   "dependencies": {
     "@alanshaw/pail": "^0.3.3",
-    "@fireproof/core": "^0.15.1",
+    "@fireproof/encrypted-blockstore": "0.16.5",
     "cross-fetch": "^4.0.0",
     "multiformats": "^12.0.1"
   }
