+++ packages/connect/src/store-remote.ts
@@ -1,7 +1,7 @@
 /* eslint-disable import/first */
 import { DownloadFnParamTypes, UploadDataFnParams } from './types'
-import type { AnyBlock, AnyLink, DbMeta } from '@fireproof/core'
-import { DataStore as DataStoreBase, MetaStore as MetaStoreBase } from '@fireproof/core'
+import type { AnyBlock, AnyLink, DbMeta } from '@fireproof/encrypted-blockstore'
+import { DataStore as DataStoreBase, MetaStore as MetaStoreBase } from '@fireproof/encrypted-blockstore'
 import { Connection } from './connection'
 import { validateDataParams, validateMetaParams } from '.'
