+++ packages/encrypted-blockstore/src/index.ts
@@ -1,2 +1,4 @@
 // import { EncryptedBlockstore, BlockstoreReader } from './transaction'
 export { EncryptedBlockstore, CompactionFetcher, BlockFetcher, CarTransaction } from './transaction'
+export { Loader } from './loader'
+export { DataStore, MetaStore } from './store'
