+++ packages/encrypted-blockstore/src/version.ts
@@ -1,1 +1,1 @@
-export const PACKAGE_VERSION = "0.16.4";
+export const PACKAGE_VERSION = "0.16.5";
