+++ packages/fireproof/src/fireproof.ts
@@ -1,5 +1,5 @@
 // ts-unused-exports:disable-next-line
 export * from './database'
-// ts-unused-exports:disable-next-line
-export * from './index'
+
+
